refactor: moved ports.types to database.types
