+++ kingdom_sdk/database/types.py
@@ -1,7 +1,9 @@
-from typing import Callable
+from typing import Callable, Union
+from uuid import UUID
 
 from sqlalchemy import MetaData, Table
 
+PrimaryKey_T = Union[int, str, UUID]
 TableFactory_T = Callable[[MetaData], Table]
 
-__all__ = ["TableFactory_T"]
+__all__ = ["PrimaryKey_T", "TableFactory_T"]
