--- kingdom_sdk/ports/types.py
@@ -1,6 +1,0 @@
-from typing import Union
-from uuid import UUID
-
-PrimaryKey_T = Union[int, str, UUID]
-
-__all__ = ["PrimaryKey_T"]
